requirements.txt
yandex_tracker_client==2.3
clickhouse_driver==0.2.*
datadog==0.44.*
apscheduler==3.9.*
requests==2.27.*
pandas==1.3.*
numpy==1.21.*
businesstime==0.3.*
holidays==0.14.*
sentry-sdk==1.6.*
python-dotenv==0.20.*